add actionTyre

+++ src/controllers/unit.controller.ts
@@ -2,6 +2,7 @@ import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
 
 const unitClient = new PrismaClient().unit;
+const prismaClient = new PrismaClient();
 
 //getAllUnit
 export const getAllUnit = async (req: Request, res: Response) => {
@@ -9,9 +10,10 @@ export const getAllUnit = async (req: Request, res: Response) => {
         const units = await unitClient.findMany({
             include: {
                 site: true,
-                tyre: true
             }
         });
+
+
 
         res.status(200).json({ data: units });
     } catch (error) {
@@ -27,7 +29,62 @@ export const getUnitById = async (req: Request, res: Response) => {
         const unit = await unitClient.findUnique({
             where: { id: unitId },
             include: {
-                site: true
+                site: true,
+                tyre1: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+                tyre2: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+                tyre3: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+                tyre4: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+                tyre5: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+                tyre6: {
+                    include: {
+                        stockTyre: {
+                            select: {
+                                serialNumber: true,
+                            }
+                        },
+                    }
+                },
+
             }
         });
 
@@ -69,25 +126,112 @@ export const getUnitBySN = async (req: Request, res: Response) => {
 //createUnit
 export const createUnit = async (req: Request, res: Response) => {
     try {
-        const { nomorUnit, hmUnit, siteId, location } = req.body;
+        const {
+            nomorUnit,
+            hmUnit,
+            kmUnit,
+            siteId,
+            location,
+            tyreIds // Array: [tyre1Id, tyre2Id, ... up to 6]
+        } = req.body;
 
-        // if (!nomor_unit || !hm_unit || !siteId) {
-        //     return res.status(400).json({ message: "nomor_unit, hm_unit, and siteId are required" });
-        // }
+        if (!nomorUnit || !siteId) {
+            res.status(400).json({ message: 'nomorUnit and siteId are required' });
+            return
+        }
 
-        const newUnit = await unitClient.create({
-            data: {
-                nomorUnit,
-                hmUnit,
-                siteId,
-                location
+        // Validasi: tidak boleh kosong atau null
+        if (!tyreIds || !Array.isArray(tyreIds) || tyreIds.length === 0) {
+            res.status(400).json({
+                message: 'tyreIds is required and cannot be empty'
+            })
+            return
+        }
+
+        if (tyreIds && tyreIds.length > 6) {
+            res.status(400).json({ message: 'You can only assign up to 6 tyres' });
+            return
+        }
+
+
+        const unit = await prismaClient.$transaction(async (tx) => {
+            // 1. Buat Unit
+            const newUnit = await tx.unit.create({
+                data: {
+                    nomorUnit,
+                    hmUnit,
+                    kmUnit,
+                    siteId,
+                    location,
+                    tyre1Id: tyreIds?.[0] ?? null,
+                    tyre2Id: tyreIds?.[1] ?? null,
+                    tyre3Id: tyreIds?.[2] ?? null,
+                    tyre4Id: tyreIds?.[3] ?? null,
+                    tyre5Id: tyreIds?.[4] ?? null,
+                    tyre6Id: tyreIds?.[5] ?? null
+                }
+            });
+
+            // 2. Update Tyre jika ada
+            if (tyreIds && tyreIds.length > 0) {
+                for (let i = 0; i < tyreIds.length; i++) {
+                    await tx.tyre.update({
+                        where: { id: tyreIds[i] },
+                        data: {
+                            isInstalled: true,
+                            isReady: false,
+                            installedUnitId: newUnit.id,
+                            positionTyre: i + 1
+                        }
+                    });
+                }
             }
+
+            for (let i = 0; i < tyreIds.length; i++) {
+                const tyreId = tyreIds[i];
+                if (!tyreId) continue;
+
+                // Update Tyre
+                const currentTyre = await tx.tyre.findUnique({
+                    where: { id: tyreId },
+                    select: {
+                        tread1: true,
+                        tread2: true
+                    }
+                });
+
+                await tx.activityTyre.create({
+                    data: {
+                        location: newUnit.location,
+                        unitId: newUnit.id,
+                        tyrePosition: i + 1,
+                        installedTyreId: tyreId,
+                        hmAtActivity: hmUnit ?? 0,
+                        kmAtActivity: kmUnit ?? 0,
+                        tread1Install: currentTyre?.tread1 ?? 0,
+                        tread2Install: currentTyre?.tread2 ?? 0,
+                        // Kosongkan tanggal, nanti diisi saat update aktivitas lanjutan
+                        dateTimeWork: null,
+                        dateTimeDone: null
+                    }
+                });
+            }
+
+
+            return newUnit;
         });
 
-        res.status(201).json({ message: "Unit created", data: newUnit });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: "Failed to create unit" });
+        res.status(201).json({ message: 'Unit created successfully', unit });
+
+    } catch (error: any) {
+        console.error('Error creating unit:', error);
+
+        if (error.code === 'P2002') {
+            res.status(409).json({ error: 'nomorUnit or tyre already assigned' });
+            return
+        }
+
+        res.status(500).json({ error: 'Internal Server Error' });
     }
 };
 
@@ -118,7 +262,10 @@ export const updateUnit = async (req: Request, res: Response) => {
 export const deleteUnit = async (req: Request, res: Response) => {
     try {
         const unitId = Number(req.params.id)
-
+        if (isNaN(unitId)) {
+            res.status(400).json({ message: "Invalid unit ID" });
+            return;
+        }
         await unitClient.delete({
             where: { id: unitId }
         });
